+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 * Added pickle load inspection [PIC100](doc/checks/PIC100.md)
 * Added django safe strings inspection [DJG102](doc/checks/DJG102.md)
+* Added hardcoded temp path read or write inspection [TMP101](doc/checks/TMP101.md)
 
 ## 1.6.0